+++ src/test/java/com/atlas/tool/impl/ListToolParameterSpecContractTest.java
@@ -57,7 +57,6 @@ class ListToolParameterSpecContractTest {
         assertListQueryContract(new CoursewareListTool(null), "courseware_list");
         assertListQueryContract(new DownloadTaskListTool(null), "download_task_list");
         assertListQueryContract(new InboxMessageListTool(null), "inbox_message_list");
-        assertListQueryContract(new MigConfigListTool(null), "mig_config_list");
         assertListQueryContract(new NamespaceListTool(null), "namespace_list");
         assertListQueryContract(new TableListTool(null), "table_list");
         assertListQueryContract(new SlurmNodeListTool(null), "slurm_node_list");